backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "PriceSentinel FastAPI backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn>=0.43.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "requests>=2.33.1",
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "joblib>=1.3.0",
    "gtts>=2.5.4",
    "langchain-groq>=1.1.2",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt==4.2.1",
]
